// src/api/types.ts
export interface CandidatItem {
  id: number
  nom: string
  prenom: string
  nomComplet: string
  initiales: string
  email: string
  telephone: string | null
  linkedin: string | null
  cvDisponible: boolean
  cvNom: string | null
  lettreDisponible: boolean
  nbCandidatures: number
  offres: string
}
// ...
export interface OffreEmploi {
  id: number
  titre: string
  departement: string | null
  typeContrat: string | null
  niveau: string | null
  mode: string | null
  statut: string
  datePublication: string
  totalCandidatures: number
  candidaturesParEtape: Record<string, number>
}

export interface Candidature {
  id: number
  offreId: number
  offreTitre: string
  candidatId: number
  candidatNom: string
  candidatPrenom: string
  nomComplet: string
  initiales: string
  email: string
  telephone: string | null
  linkedin: string | null
  etape: 'BOITE_RECEPTION' | 'BROUILLON' | 'ENTRETIEN_TEL' | 'ENTRETIEN_PHYSIQUE' | 'EMBAUCHE'
  dateEntretien: string | null
  dateCreation: string
  cvDisponible: boolean
  cvNom: string | null
  lettreDisponible: boolean
  historique: string[]
}